fixed another compile warning

+++ src/pyglue/PyUtil.cpp
@@ -310,11 +310,11 @@ OCIO_NAMESPACE_ENTER
         {
             if(PyList_Check(pyobj))
             {
-                return PyList_GET_SIZE(pyobj);
+                return static_cast<int>(PyList_GET_SIZE(pyobj));
             }
             else if(PyTuple_Check(pyobj))
             {
-                return PyTuple_GET_SIZE(pyobj);
+                return static_cast<int>(PyTuple_GET_SIZE(pyobj));
             }
             return -1;
         }
